Create doc.go
Testing automatic godoc reference.

In theory this would be our main godoc package overview documentation. 

Each additional file (and it's functions) would have it's comments parsed, and automatically documented under this package.

Not sure if this generating the godoc would need to be added to the build process (via the go doc or go tool commands) or how exactly that process happens. It will be automatically updated and maintained however somehow.